+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for stripes-template-editor
 
+## IN PROGRESS
+
+* Should have possible handle disabled state for loop by passed function. Refs STRIPES-856.
+
 ## [3.2.0](https://github.com/folio-org/stripes-template-editor/tree/v3.2.0) (2023-03-03)
 [Full Changelog](https://github.com/folio-org/stripes-template-editor/compare/v3.1.1...v3.2.0)